Add Playwright Step Details "Elements" tab (#10074)
* Extract Playwright and Cypress step details into separate files

* Pass test event directly to details components

* Move step details components into a subfolder

* Add "Elements" and "Source" tabs to the Playwright Step Details panel
