+++ src/ui/components/TestSuite/views/TestRecording/TestEventStepDetails/TestEventDetails.module.css
@@ -97,3 +97,40 @@
 .SourceCodeLineSelected {
   background-color: var(--background-color-current-execution-point);
 }
+
+.TabsContainer {
+  display: flex;
+  flex-direction: column;
+  overflow: auto;
+}
+
+.Tabs {
+  display: flex;
+  background: var(--tab-bgcolor);
+  border-bottom: 1px solid var(--theme-splitter-color);
+  flex: 0 0 auto;
+}
+
+.Tab {
+  padding: 0.5rem;
+  border-bottom-color: transparent;
+  background-color: var(--tab-bgcolor);
+  color: var(--tab-color);
+}
+.Tab[data-active] {
+  background-color: var(--tab-selected-bgcolor);
+  color: var(--tab-selected-color);
+  cursor: default;
+}
+.Tab:hover:not(.Tab[data-active]) {
+  background-color: var(--tab-hover-bgcolor);
+  color: var(--tab-hover-color);
+}
+.Tab:focus {
+  outline: none;
+}
+
+.Panel {
+  height: 100%;
+  flex: 1 1 auto;
+}
